повыносила с роутеров в отдельные контролеры и передала в роутеры
