+++ controllers/contacts/removeContact.js
@@ -1,0 +1,23 @@
+const contactsOperations = require("../../models/contacts");
+const createError = require("http-errors");
+const removeContact = async (req, res, next) => {
+    try {
+        const {contactId} = req.params;
+        const result = await contactsOperations.removeContact(contactId);
+        if (!result) {
+            throw createError(404, "Not found");
+        }
+        res.json({
+            status: "success",
+            code: 200,
+            message: "contact deleted",
+            data: {
+                result
+            }
+        })
+    } catch (error) {
+        next(error);
+    }
+}
+
+module.exports = removeContact;
